requirements.txt
asgiref==3.7.2
Django==5.0.1
fire==0.5.0
fonttools==4.47.2
fpdf==1.7.2
lxml==5.1.0
numpy==1.26.3
opencv-python==4.9.0.80
pdf2docx==0.5.7
PyMuPDF==1.23.11
PyMuPDFb==1.23.9
python-docx==1.1.0
six==1.16.0
sqlparse==0.4.4
termcolor==2.4.0
typing_extensions==4.9.0
tzdata==2023.4
